pyproject.toml
[project]
name = "codesphere"

version = "0.4.0"
description = "Use Codesphere within python scripts."
readme = "README.md"
license = { file="LICENSE" }
authors = [
    { name = "Datata1", email = "jan-david.wiederstein@codesphere.com" }
]
requires-python = ">=3.12.9"
dependencies = [
    "aiohttp>=3.12.13",
    "aiohttp-retry>=2.9.1",
    "httpx>=0.28.1",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.11.0",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
    "typing-extensions>=4.14.0",
    "urllib3>=2.4.0",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License", 
    "Operating System :: OS Independent",
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/codesphere"]

[project.optional-dependencies]
dev = [
    "commitizen>=4.8.3",
    "pre-commit>=4.2.0",
    "pytest>=8.4.0",
    "pytest-asyncio",
    "pytest-cov>=6.2.1",
    "ruff>=0.11.13",
]

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "v$version"
version_scheme = "pep440"
version_provider = "uv"
update_changelog_on_bump = true
major_version_zero = true
style = [
    ["qmark", "fg:#ff9d00 bold"],
    ["question", "bold"],
    ["answer", "fg:#ff9d00 bold"],
    ["pointer", "fg:#ff9d00 bold"],
    ["highlighted", "fg:#ff9d00 bold"],
    ["selected", "fg:#cc5454"],
    ["separator", "fg:#cc5454"],
    ["instruction", ""],
    ["text", ""],
    ["disabled", "fg:#858585 italic"]
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = [
  "."
]
python_files = "test_*.py *_test.py"
python_functions = "test_*"
python_classes = "Test*"
asyncio_mode = "auto"
markers = [
    "integration: mark test as integration test (requires API token and --run-integration flag)",
]

[tool.coverage.run]
source = ["api", "handler", "tasks"]
branch = true
omit = [
    ".github/*",
    ".ruff_cache/*",
    "docs/*",
    "examples/**",
    "scripts/**",
    "*/__main__.py",
    "tests/*",
    ".venv/*",
    "__init__.py"
]

[tool.coverage.report]
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
    "@(abc\\.)?abstractmethod",
    "def __repr__",
    "def __str__",
]
include_namespace_packages = true

[tool.coverage.html]
directory = "test-results"

[tool.bandit]
targets = ["src/codesphere_sdk/__init__.py"]  # add here folders to check
exclude_dirs = ["tests", ".venv", ".uv", "examples", "docs", ".github", "scripts"]
skips = ["B101"]

[project.urls]
Homepage = "https://codesphere.com/"
Repository = "https://github.com/Datata1/codesphere-python"
"Bug Tracker" = "https://github.com/Datata1/codesphere-python/issues"
